refactor(stella-cli): extract the Git diagnostic argv mapping into its own module
The `UntrackedPatch` arm pushed `agent/tools.rs` to 1504 lines, over the
1500-line limit, and that file is not grandfathered — so the arm had to
land somewhere else rather than the ceiling move.

`impl DiagnosticRunner for GitDiagnosticRunner` is the natural seam: it is
the one place a `DiagnosticInvocation` becomes a process, and it needs
nothing from `tools` but the runner's root, its baseline commit, and the
two spawn helpers. Moving it takes tools.rs to 1470.

`agent.rs` grows by the one line that declares the module — the
irreducible case the baseline documents an escape hatch for, so the +1
lands as a visible baseline diff. Regenerating also ratchets
command_deck.rs (-55) and bus.rs (-235) down to what main already earned.
